Two click update. Clears between shots.
Fixed the setInterval so that it has a clearInterval that fires after
the line has drawn. Now that there is a clearInterval, it will only
draw one line at a time.

BUG: If you try to draw lines really rapidly, it will start a new loop
before the clearInterval of the previous loop has been called,
therefore creating an endless loop of the first one while you go on
drawing new lines. Fix coming soon.