Laser now erases tail
By simply drawing a black line right after the first line has been
drawn, it simulates what a laser would look like.

This will NOT be a good option if it is to be drawn over a scene in
which the background color is not constant. (But if our inner grid/map
is a single color, this would be a great option).